pyproject.toml
[project]
name = "task-registry"
version = "0.1.0"
description = "Registry and schema management for VLMHyperBench tasks, datasets, and metrics."
authors = [
    {name = "Anton Shiryaev", email = "med.phisiker@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[build-system]
requires = ["uv_build>=0.9.26"]
build-backend = "uv_build"

[tool.uv]
package = true